refactoring to uniqueness checker
