+++ components/uniqueness/uniqueness-checker-impl-osgi-tests/src/integrationTest/kotlin/net/corda/uniqueness/checker/impl/osgitests/MessageBusIntegrationTests.kt
@@ -280,7 +280,6 @@ class MessageBusIntegrationTests {
 
         uniquenessCheckerLifecycle = BatchedUniquenessCheckerLifecycleImpl(
             lifecycleCoordinatorFactory,
-            configurationReadService,
             subscriptionFactory,
             externalEventResponseFactory,
             backingStore,
