+++ components/uniqueness/uniqueness-checker-impl/src/main/kotlin/net/corda/uniqueness/checker/impl/BatchedUniquenessCheckerLifecycleImpl.kt
@@ -1,12 +1,6 @@
 package net.corda.uniqueness.checker.impl
 
-import net.corda.configuration.read.ConfigChangedEvent
-import net.corda.configuration.read.ConfigurationReadService
-import net.corda.data.flow.event.FlowEvent
-import net.corda.data.uniqueness.UniquenessCheckRequestAvro
 import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
-import net.corda.libs.configuration.SmartConfig
-import net.corda.libs.configuration.helper.getConfig
 import net.corda.lifecycle.DependentComponents
 import net.corda.lifecycle.LifecycleCoordinator
 import net.corda.lifecycle.LifecycleCoordinatorFactory
@@ -17,11 +11,8 @@ import net.corda.lifecycle.StartEvent
 import net.corda.lifecycle.StopEvent
 import net.corda.lifecycle.createCoordinator
 import net.corda.messaging.api.constants.WorkerRPCPaths.UNIQUENESS_PATH
-import net.corda.messaging.api.subscription.config.SubscriptionConfig
 import net.corda.messaging.api.subscription.config.SyncRPCConfig
 import net.corda.messaging.api.subscription.factory.SubscriptionFactory
-import net.corda.schema.Schemas
-import net.corda.schema.configuration.ConfigKeys.MESSAGING_CONFIG
 import net.corda.uniqueness.backingstore.BackingStoreLifecycle
 import net.corda.uniqueness.checker.UniquenessChecker
 import net.corda.uniqueness.checker.UniquenessCheckerLifecycle
@@ -38,8 +29,6 @@ import org.slf4j.LoggerFactory
 class BatchedUniquenessCheckerLifecycleImpl @Activate constructor(
     @Reference(service = LifecycleCoordinatorFactory::class)
     coordinatorFactory: LifecycleCoordinatorFactory,
-    @Reference(service = ConfigurationReadService::class)
-    private val configurationReadService: ConfigurationReadService,
     @Reference(service = SubscriptionFactory::class)
     private val subscriptionFactory: SubscriptionFactory,
     @Reference(service = ExternalEventResponseFactory::class)
@@ -50,10 +39,8 @@ class BatchedUniquenessCheckerLifecycleImpl @Activate constructor(
     uniquenessChecker: UniquenessChecker
 ) : UniquenessCheckerLifecycle, UniquenessChecker by uniquenessChecker {
     private companion object {
-        const val GROUP_NAME = "uniqueness.checker"
         const val CONFIG_HANDLE = "CONFIG_HANDLE"
         const val SUBSCRIPTION_NAME = "Uniqueness Check"
-        const val SUBSCRIPTION = "SUBSCRIPTION"
         const val RPC_SUBSCRIPTION = "RPC_SUBSCRIPTION"
 
         private val log = LoggerFactory.getLogger(this::class.java.enclosingClass)
@@ -80,36 +67,26 @@ class BatchedUniquenessCheckerLifecycleImpl @Activate constructor(
     }
 
     private fun eventHandler(event: LifecycleEvent, coordinator: LifecycleCoordinator) {
-        log.info("Uniqueness checker received event $event")
+        log.trace("Uniqueness checker received event $event")
         when (event) {
             is StartEvent -> {
-                configurationReadService.start()
                 dependentComponents.registerAndStartAll(coordinator)
+            }
+            is RegistrationStatusChangeEvent -> {
+                if (event.status == LifecycleStatus.UP) {
+                    log.trace("The status of event: $event changed to ${event.status}, starting subscription.")
+                    initialiseRpcSubscription()
+                    coordinator.updateStatus(LifecycleStatus.UP)
+                } else {
+                    coordinator.updateStatus(event.status)
+                    coordinator.closeManagedResources(setOf(CONFIG_HANDLE))
+                    log.trace("The status of event: $event changed to ${event.status}, stopping subscription.")
+                }
             }
             is StopEvent -> {
                 dependentComponents.stopAll()
-            }
-            is RegistrationStatusChangeEvent -> {
-                log.info("Uniqueness checker is ${event.status}")
-
-                if (event.status == LifecycleStatus.UP) {
-                    initialiseRpcSubscription()
-
-                    coordinator.createManagedResource(CONFIG_HANDLE) {
-                        configurationReadService.registerComponentForUpdates(
-                            coordinator,
-                            setOf(MESSAGING_CONFIG)
-                        )
-                    }
-                } else {
-                    coordinator.closeManagedResources(setOf(CONFIG_HANDLE))
-                }
-                coordinator.updateStatus(event.status)
-            }
-            is ConfigChangedEvent -> {
-                log.info("Received configuration change event, (re)initialising subscription")
-                initialiseSubscription(event.config.getConfig(MESSAGING_CONFIG))
-                // RPC Subscription doesn't need to be re-created because it doesn't take config.
+                coordinator.closeManagedResources(setOf(RPC_SUBSCRIPTION))
+                log.trace("Received stop event: $event, stopping subscription.")
             }
             else -> {
                 log.warn("Unexpected event ${event}, ignoring")
@@ -117,28 +94,10 @@ class BatchedUniquenessCheckerLifecycleImpl @Activate constructor(
         }
     }
 
-    private fun initialiseSubscription(config: SmartConfig) {
-        lifecycleCoordinator.createManagedResource(SUBSCRIPTION) {
-            subscriptionFactory.createDurableSubscription(
-                SubscriptionConfig(GROUP_NAME, Schemas.UniquenessChecker.UNIQUENESS_CHECK_TOPIC),
-                UniquenessCheckMessageProcessor(
-                    this,
-                    externalEventResponseFactory
-                ),
-                config,
-                null
-            ).also {
-                it.start()
-            }
-        }
-    }
-
     private fun initialiseRpcSubscription() {
-        val processor = UniquenessCheckRpcMessageProcessor(
+        val processor = UniquenessCheckMessageProcessor(
             this,
-            externalEventResponseFactory,
-            UniquenessCheckRequestAvro::class.java,
-            FlowEvent::class.java
+            externalEventResponseFactory
         )
         lifecycleCoordinator.createManagedResource(RPC_SUBSCRIPTION) {
             val rpcConfig = SyncRPCConfig(SUBSCRIPTION_NAME, UNIQUENESS_PATH)
