+++ components/uniqueness/uniqueness-checker-impl/src/main/kotlin/net/corda/uniqueness/checker/impl/UniquenessCheckMessageProcessor.kt
@@ -1,66 +1,35 @@
 package net.corda.uniqueness.checker.impl
 
+import net.corda.data.flow.event.FlowEvent
 import net.corda.data.uniqueness.UniquenessCheckRequestAvro
 import net.corda.data.uniqueness.UniquenessCheckResultUnhandledExceptionAvro
 import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
-import net.corda.messaging.api.processor.DurableProcessor
-import net.corda.messaging.api.records.Record
-import net.corda.tracing.BatchRecordTracer
-import net.corda.tracing.traceBatch
+import net.corda.messaging.api.processor.SyncRPCProcessor
 import net.corda.uniqueness.checker.UniquenessChecker
 
 /**
- * Processes messages received from the uniqueness check topic, and responds using the external
+ * Processes messages received from the RPC calls, and responds using the external
  * events response API.
  */
 class UniquenessCheckMessageProcessor(
     private val uniquenessChecker: UniquenessChecker,
-    private val externalEventResponseFactory: ExternalEventResponseFactory
-) : DurableProcessor<String, UniquenessCheckRequestAvro> {
+    private val externalEventResponseFactory: ExternalEventResponseFactory,
+) : SyncRPCProcessor<UniquenessCheckRequestAvro, FlowEvent> {
 
-    override val keyClass = String::class.java
-    override val valueClass = UniquenessCheckRequestAvro::class.java
+    override val requestClass = UniquenessCheckRequestAvro::class.java
+    override val responseClass = FlowEvent::class.java
 
-    override fun onNext(events: List<Record<String, UniquenessCheckRequestAvro>>): List<Record<*, *>> {
-
-        val batchTracer = createBatchTracer(events)
-
-        val requests = events.mapNotNull { it.value }
-
-        return uniquenessChecker.processRequests(requests).map { (request, response) ->
+    override fun process(request: UniquenessCheckRequestAvro): FlowEvent {
+        return uniquenessChecker.processRequests(listOf(request)).map { (request, response) ->
             if (response.result is UniquenessCheckResultUnhandledExceptionAvro) {
-                batchTracer.error(
-                    request,
-                    externalEventResponseFactory.platformError(
-                        request.flowExternalEventContext,
-                        (response.result as UniquenessCheckResultUnhandledExceptionAvro).exception
-                    )
+                externalEventResponseFactory.platformError(
+                    request.flowExternalEventContext,
+                    (response.result as UniquenessCheckResultUnhandledExceptionAvro).exception
                 )
             } else {
-                batchTracer.complete(
-                    request,
-                    externalEventResponseFactory.success(request.flowExternalEventContext, response)
-                )
+                externalEventResponseFactory.success(request.flowExternalEventContext, response)
             }
-        }
+        }.single().value!!
     }
 }
 
-private fun BatchRecordTracer.error(request: UniquenessCheckRequestAvro, record: Record<*, *>): Record<*, *> {
-    return request.flowExternalEventContext?.requestId?.let { id -> this.completeSpanFor(id, record) } ?: record
-}
-
-private fun BatchRecordTracer.complete(request: UniquenessCheckRequestAvro, record: Record<*, *>): Record<*, *> {
-    return request.flowExternalEventContext?.requestId?.let { id -> this.completeSpanFor(id, record) } ?: record
-}
-
-private fun createBatchTracer(events: List<Record<String, UniquenessCheckRequestAvro>>): BatchRecordTracer {
-    return traceBatch("Uniqueness Check Request").apply {
-        events.forEach { event ->
-            val id = event.value?.flowExternalEventContext?.requestId
-            if (id != null) {
-                this.startSpanFor(event, id)
-            }
-        }
-    }
-}
